+++ packages/fleury_widgets/lib/src/date_picker.dart
@@ -126,7 +126,7 @@ class _DatePickerState extends State<DatePicker> implements TextInputClaimant {
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    Focus.maybeOf(context); // rebuild on focus change
+    FocusManager.maybeOf(context); // rebuild on focus change
     final registration = FormControlScope.maybeOf(context);
     if (!identical(registration, _formRegistration)) {
       _formRegistration?.release(this);